requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
asgiref==3.11.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.1
captcha==0.7.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cmake==4.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
decorator==5.2.1
deepface==0.0.97
Django==5.2.10
extra-streamlit-components==0.1.81
filelock==3.20.3
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2026.1.0
future==1.0.0
gast==0.7.0
gdown==5.2.1
geocoder==1.38.1
geographiclib==2.1
geoip2==5.2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.76.0
gunicorn==23.0.0
h5py==3.14.0
idna==3.11
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
lightecc==0.0.4
lightphe==0.0.20
lz4==4.4.5
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
maxminddb==3.0.0
mdurl==0.1.2
mediapipe==0.10.14
ml_dtypes==0.5.3
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.7.0
mysql-connector-python==8.0.28
namex==0.1.0
narwhals==2.10.1
networkx==3.4.2
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.13.0.90
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
protobuf==6.33.4
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.2
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
ratelim==0.1.6
referencing==0.37.0
requests==2.32.5
retina-face==0.0.17
rich==14.2.0
rpds-py==0.28.0
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
smmap==5.0.2
sounddevice==0.5.3
soupsieve==2.8.3
SQLAlchemy==2.0.44
sqlmodel==0.0.27
sqlparse==0.5.5
st-pages==1.0.1
streamlit==1.51.0
streamlit-authenticator==0.4.2
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.21.0
termcolor==3.3.0
tf-slim==1.1.0
tf_keras==2.20.1
threadpoolctl==3.6.0
toml==0.10.2
torch==2.11.0
torchvision==0.26.0
tornado==6.5.2
tqdm==4.67.1
twilio==9.9.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.5
wrapt==2.0.1
yarl==1.22.0
